+++ tests/test_web.py
@@ -211,7 +211,7 @@ def test_tables_statistics_version(
 ) -> None:
     table = sample_config["tables"][0]
 
-    response = client.get(f"/tables/{table['name']}/statistics?version=0")
+    response = client.get(f"/tables/{table['name']}/statistics", params={"version": 0})
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -262,7 +262,9 @@ def test_tables_view_limit(
     table = sample_config["tables"][0]
     selected_limit = 1
 
-    response = client.get(f"/tables/{table['name']}/view?limit={selected_limit}")
+    response = client.get(
+        f"/tables/{table['name']}/view", params={"limit": selected_limit}
+    )
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -282,9 +284,11 @@ def test_tables_view_cols(
         delta_table.schema().fields[i].name for i in range(num_fields - 1)
     ]
     filtered_columns = [delta_table.schema().fields[num_fields - 1].name]
-    qs = "&".join(f"cols={col}" for col in selected_columns)
 
-    response = client.get(f"/tables/{table['name']}/view?{qs}")
+    response = client.get(
+        f"/tables/{table['name']}/view",
+        params=[("cols", col) for col in selected_columns],
+    )
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -315,7 +319,9 @@ def test_table_view_sort_asc(
     default_limit = 10
     sort_column = "temperature"
 
-    response = client.get(f"/tables/{table['name']}/view?sort_asc={sort_column}")
+    response = client.get(
+        f"/tables/{table['name']}/view", params={"sort_asc": sort_column}
+    )
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -340,7 +346,9 @@ def test_table_view_sort_desc(
     default_limit = 10
     sort_column = "temperature"
 
-    response = client.get(f"/tables/{table['name']}/view?sort_desc={sort_column}")
+    response = client.get(
+        f"/tables/{table['name']}/view", params={"sort_desc": sort_column}
+    )
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -362,7 +370,7 @@ def test_tables_view_version(
     table = sample_config["tables"][0]
     default_limit = 10
 
-    response = client.get(f"/tables/{table['name']}/view?version=0")
+    response = client.get(f"/tables/{table['name']}/view", params={"version": 0})
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -394,7 +402,7 @@ def test_tables_query(
     selected_limit = 1
     sql_query = f"select {selected_column} from {sample_config['tables'][0]['name']} limit {selected_limit}"
 
-    response = client.get(f"/tables/query?sql={sql_query}")
+    response = client.get("/tables/query", params={"sql": sql_query})
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -422,12 +430,40 @@ def test_tables_query(
     assert not all(str(row) in all_td for row in df[selected_column][selected_limit:])
 
 
+@pytest.mark.parametrize(
+    ("start_date", "end_date"),
+    [("", ""), ("2025-01-01", ""), ("", "2025-01-31"), ("2025-01-01", "2025-01-31")],
+)
+def test_tables_query_parameters(
+    client: TestClient,
+    sample_config: dict[str, Any],
+    start_date: str,
+    end_date: str,
+) -> None:
+    sql_query = f"select * from {sample_config['tables'][0]['name']} where day between $start_date and $end_date"
+
+    response = client.get(
+        "/tables/query",
+        params={"sql": sql_query, "start_date": start_date, "end_date": end_date},
+    )
+    assert response.status_code == HTTPStatus.OK
+
+    html = response.content.decode()
+    soup = BeautifulSoup(html, "html.parser")
+
+    assert soup.find("h3", string="Parameters")
+    assert soup.find("label", string="start_date")
+    assert soup.find("input", attrs={"name": "start_date", "value": start_date})
+    assert soup.find("label", string="end_date")
+    assert soup.find("input", attrs={"name": "end_date", "value": end_date})
+
+
 def test_tables_query_invalid(
     client: TestClient, delta_table: deltalake.DeltaTable
 ) -> None:
     sql_query = "select * from unknown_table"
 
-    response = client.get(f"/tables/query?sql={sql_query}")
+    response = client.get("/tables/query", params={"sql": sql_query})
     assert response.status_code == HTTPStatus.OK
 
     html = response.content.decode()
@@ -676,6 +712,53 @@ def test_queries_view(client: TestClient, sample_config: dict[str, Any]) -> None
     assert all(col in all_th for col in {"day", "avg_temperature"})
 
 
+def test_queries_view_parameters_default(
+    client: TestClient, sample_config: dict[str, Any]
+) -> None:
+    query = sample_config["queries"][1]
+
+    response = client.get(f"/queries/{query['name']}/view")
+    assert response.status_code == HTTPStatus.OK
+    assert response.history[0].has_redirect_location
+
+    html = response.content.decode()
+    soup = BeautifulSoup(html, "html.parser")
+
+    assert soup.find("h2", string=query["title"])
+    assert soup.find("h3", string="Parameters")
+
+    for param_name, param_data in query["parameters"].items():
+        assert soup.find("label", string=param_name)
+        assert soup.find(
+            "input", attrs={"name": param_name, "value": param_data["default"]}
+        )
+
+
+def test_queries_view_parameters(
+    client: TestClient, sample_config: dict[str, Any]
+) -> None:
+    query = sample_config["queries"][1]
+
+    response = client.get(
+        f"/queries/{query['name']}/view",
+        params={k: v["default"] for k, v in query["parameters"].items()},
+    )
+    assert response.status_code == HTTPStatus.OK
+    assert len(response.history) == 0
+
+    html = response.content.decode()
+    soup = BeautifulSoup(html, "html.parser")
+
+    assert soup.find("h2", string=query["title"])
+    assert soup.find("h3", string="Parameters")
+
+    for param_name, param_data in query["parameters"].items():
+        assert soup.find("label", string=param_name)
+        assert soup.find(
+            "input", attrs={"name": param_name, "value": param_data["default"]}
+        )
+
+
 def test_queries_view_invalid(
     monkeypatch: pytest.MonkeyPatch,
     sample_config: dict[str, Any],
@@ -713,7 +796,7 @@ def test_tables_query_export_csv(
     selected_limit = 2
     sql_query = f"select {selected_column} from {sample_config['tables'][0]['name']} limit {selected_limit}"
 
-    response = client.get(f"/tables/query/csv?sql={sql_query}")
+    response = client.get("/tables/query/csv", params={"sql": sql_query})
     assert response.status_code == HTTPStatus.OK
     assert response.headers["content-type"] == "text/csv; charset=utf-8"
     assert "attachment" in response.headers["content-disposition"]
@@ -730,7 +813,7 @@ def test_queries_view_export_csv(
 ) -> None:
     query = sample_config["queries"][0]
 
-    response = client.get(f"/tables/query/csv?sql={query['sql']}")
+    response = client.get("/tables/query/csv", params={"sql": query["sql"]})
     assert response.status_code == HTTPStatus.OK
     assert response.headers["content-type"] == "text/csv; charset=utf-8"
     assert "attachment" in response.headers["content-disposition"]