+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - web: display app version in sidebar
 - web: use CodeMirror SQL query editor
+- web: add support for tables query parameters
+- web: add support for predefined queries parameters
+
+### Changed
+- demo: add parameters to daily avg temperature query
 
 ### Fixed
 - handle empty SQL queries